fix everything

+++ src/news.php
@@ -1,6 +1,8 @@
 <?php
 include 'php/session_check.php';
 include 'php/db_connect.php';
+
+$is_admin = isset($_SESSION['user_role']) && $_SESSION['user_role'] === 'Admin';
 ?>
 
 <!DOCTYPE html>
@@ -16,28 +18,9 @@ include 'php/db_connect.php';
         <h1>REPUBLICKI IZBORI REPUBLIKE SRBIJE</h1>
     </div>
 
-    <div id="navbar-container">
-        <nav>
-            <uL>
-                <li><a href="first_nav.html">Naslovna</a></li>
-                <li>
-                    <a href="results.html">Izborni rezultati</a>
-                    <ul>
-                        <li><a href="#">Unesi rezultate</a></li>
-                        <li><a href="#">Azuriraj rezultate</a></li>
-                    </ul>
-                </li>
-                <li>
-                    <a href="news.html">Vesti</a>
-                    <ul>
-                        <li><a href="add_news.html">Unesi vest</a></li>
-                        <li><a href="news_input.html">Azuriraj vesti</a></li>
-                    </ul>
-                </li>
-                <li><a href="login.html">Uloguj se</a></li>
-            </uL>
-        </nav>
-    </div>
+    <?php
+        include('navigation.php');
+    ?>
 
     <div id="news-page-container">
             <ul id="news-list-element">
@@ -52,8 +35,10 @@ include 'php/db_connect.php';
                         echo "<h5><strong>" . $row['Naslov'] . "</strong></h5>";
                         echo "<p>Datum: " . $row['Datum'] . "</p>";
                         echo "<p>" . $row['Sadrzaj'] . "</p>";
-                        echo "<button onClick=\"deleteNews('" . $row['Naslov'] . "')\">Izbrisi Vest</button>";
-                        echo "<button onClick=\"adjustNews('" . $row['Naslov'] . "')\">Azuriraj Vest</button>";
+                        if ($is_admin) {
+                            echo "<button onClick=\"deleteNews('" . $row['Naslov'] . "')\">Izbrisi Vest</button>";
+                            echo "<button onClick=\"adjustNews('" . $row['Naslov'] . "')\">Azuriraj Vest</button>";
+                        }
                         echo "</li>";
                     }
                 } else {
